Rework from movies to books and add CSV file for books
BAsically copied the movie files and changed them to books, made a short
test with System.out.prints to see that i could chain them as I did with
the movies before. Found an CSV-file which I could use that is only 1
file instead of 3, which makes the project easier.